Refactored modules to all lower case
