+++ destiny/__init__.py
@@ -2,8 +2,8 @@
 # This is for easily exposing visible names so they can be called like via
 # destiny.CarnageReport() instead of destiny.CarnageReport.CarnageReport()
 
-from .Player import Player
-from .Game import Game
-from .Guardian import Guardian
-from .ReportContext import ReportContext
-from .Manifest import Manifest
+from .player import Player
+from .game import Game
+from .guardian import Guardian
+from .report import ReportContext
+from .manifest import Manifest
